Fall back to Text::LineFold if Text::WrapI18N is not available
The main reason for this change is that installing
Text::WrapI18N on some systems is impossible -- in
particular, Android's libc doesn't provide the functions
that the module needs to work.

Unfortunately, that renders MooX::Options entirely
inoperable.  Fortunately, Text::LineFold does not
suffer from this limitation and, as a bonus, for UTF-8
encoded text it usually wraps more accurately than
Text::WrapI18N, since it is really just a thin wrapper
around Unicode::LineBreak.