I18n using Locale::TextDomain
Localization encompasses:

  - help messages
  - parameter type keywords
  - error messages

German translation files supplied for locales `de` and
`de_DE.UTF-8`.

The Locale::TextDomain module is loaded on demand.

Issues

The localization is incomplete, as some messages are generated by
dependent modules (for a discussion on this point see [1,2]).

The localization is naturally incomplete, as currently only german
language files are supplied.

[1]: https://github.com/celogeek/MooX-Options/pull/38
[2]: http://stackoverflow.com/questions/29749855/how-to-localize-cpan-module-and-dependencies