Adjust for Role-Tiny 2.001003: test suite and namespace::clean compatibility documentation
Upgrading `Role-Tiny` from 2.001001 to 2.001003 breaks a test case:

	t/16-namespace_clean.t ......................
	ok 1 - TestNamespaceClean is a package
	Can't locate object method "__" via package "TestNamespaceClean" at /usr/share/perl5/MooX/Options/Role.pm line 352.

One way to fix this is to ensure `namespace::clean` does not hide the methods
exported by `MooX::Locale::Passthrough`, as this commit does. But this problem
might be the symptom of a regression in `Role-Tiny`, that perhaps we should not
just paper over.